test: release devices before each test ci: move ios integration test to its own folder as it didn't get picked up locally
